file and folder name changed, ApiResource emplacement changed. Preparations for resource post operation, not working anymore
