+++ src/Processor/UserProcessor.php
@@ -6,8 +6,8 @@ namespace App\Processor;
 
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProcessorInterface;
-use App\ApiResource\UserApi;
 use App\Domain\User\UserCreationOrUpdate;
+use App\DTO\CreateUser;
 use App\Repository\UserRepository;
 
 readonly class UserProcessor implements ProcessorInterface
@@ -20,7 +20,7 @@ readonly class UserProcessor implements ProcessorInterface
 
     public function process(mixed $data, Operation $operation, array $uriVariables = [], array $context = []): void
     {
-        if ($data instanceof UserApi) {
+        if ($data instanceof CreateUser) {
             $emailHash = hash('sha256', $data->getEmail());
             $username = $data->getUserName();
